mobilenetv3-vector-service/requirements.txt
tensorflow==2.10.1
numpy>=1.20,<1.24
pillow>=10.2.0
requests==2.31.0
